requirements.txt
absl-py
accelerate==0.31.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==3.7.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
bidict==0.23.1
bitsandbytes==0.43.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
chainlit==1.1.306
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
comm==0.2.2
ctransformers==0.2.27
dataclasses-json==0.5.14
datasets==2.20.0
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.3
fastapi-cli==0.0.4
filelock==3.14.0
filetype==1.2.0
flatbuffers==24.3.25
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.5.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
googleapis-common-protos==1.63.2
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
importlib_metadata==8.0.0
inquirerpy==0.3.4
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpickle==3.2.2
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.3.3
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.24
langchain-text-splitters==0.2.1
langsmith==0.1.76
Lazify==0.4.0
libclang==18.1.1
literalai==0.0.607
llama_cpp_python==0.2.83
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
ollama==0.3.0
openai==1.37.1
openpyxl==3.1.5
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.4
packaging==23.2
pandas==2.2.2
parso==0.8.4
peft==0.11.1
pexpect==4.9.0
pfzy==0.3.4
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
pydot==2.0.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
pytz==2024.1
pyvis==0.3.2
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
robotframework==7.0.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
streamlit==1.37.0
sympy==1.12.1
syncer==2.0.3
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
threadpoolctl==3.5.0
together==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.3
triton==2.3.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uptrace==1.26.0
urllib3==2.2.1
uvicorn==0.25.0
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.20.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
